ADEN-5573 moved all VUAP tests to TestAdsVuapOasis.java
